UPSTREAM: <carry>: feat: mount EmptyDir volumes for launcher write locations
Launcher writes input artifacts to root paths /gcs, /minio, and /s3.
These paths are not accessible by non-root users by default, which is
problematic in locked-down Kubernetes installations and/or OpenShift.
/gcs is currently a contract for KFP v2 python component wrappers, so
the path cannot be changed.

Mount an EmptyDir scratch volume to these paths to work around this.

Additionally, /.local and /.cache are written to by pip, so add
EmptyDir mounts for those too.

Fixes: https://issues.redhat.com/browse/RHOAIENG-1889

Ref: https://github.com/kubeflow/pipelines/issues/5673
Ref: https://github.com/kubeflow/pipelines/issues/7345