UPSTREAM: <carry>: add artifact storage api for backend
This change introduces a new api for artifact data retrieval for mlpipeline api server, to be consumed by UI and end users. Two new endpoints are introduced, allowing for 1) collection retrieval and 2) individual artifact metadata retrieval.

When an individual artifact resource is retrieved, users/ui can specify a share_url option that provides a pre-signed url that with a very short expiry date that can be used as an artifact download link directly from the originating object store provider. This is currently only implemented for s3-compliant object store backends.